Door Interacts
- B2 Door Data assets are test files. Please remove once replaced
- Doors are now interactable. Added a new manager in M3 called Door Popup Manager to handle this, along with a new Scriptable Object Door Data.
- Big Door script now inherits from Door so it can benefit from the door interactable changes (and because it makes sense like that)
- Modified CollectibleController to account for the edge cases of Door not sending notifications. THIS CAN BREAK SOMETHING IF ONCLAIMRESULT IS MORE THAN NOTIFICATIONS
- Changed Report Manager to be usable by Door (Hijack typing system)
- Changed collectible manager so door could sidestep adding items
- Removed unused files - PType, my M3,
- Moved files around - BigGuy Melee to Enemy, AnimationEventHandler to Unused,
