+++ Assets/B2 Door (1).asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a9031d5d3b2c02488b424603664e6f3, type: 3}
+  m_Name: B2 Door (1)
+  m_EditorClassIdentifier: 
+  sprite: {fileID: 21300000, guid: 11ec875c4f1830643a2be1a108eddcad, type: 3}
+  text: This is the second fabulous door.
