+++ Assets/B2 Door (1).asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 830f2a2a1b3546f498948515a70c5cfc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
