+++ Assets/B2DoorEpic.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a9031d5d3b2c02488b424603664e6f3, type: 3}
+  m_Name: B2DoorEpic
+  m_EditorClassIdentifier: 
+  sprite: {fileID: 0}
+  text: This is a fabulous door.
