+++ Assets/B2DoorEpic.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 15e0246c0530a3b4e91033a2ee317298
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
