+++ Assets/BigDoor.cs
@@ -2,14 +2,12 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class BigDoor : MonoBehaviour
+public class BigDoor : Door
 {
     private Animator animator;
 
     [SerializeField] private GameObject exitSign;
     private GameObject signLight;
-
-    public bool isOpen;
 
     private void Awake() {
         animator = GetComponent<Animator>();
@@ -17,26 +15,10 @@ public class BigDoor : MonoBehaviour
         signLight = exitSign.transform.GetChild(0).gameObject;
     }
 
-    public void OpenDoor() {
+    public override void OpenDoor() {
         animator.SetBool("IsUnlocked", true);
         isOpen = true;
         AudioControl.Instance.PlaySFX("Door Opening", gameObject, 0.2f, 1f);
         signLight.GetComponent<UnityEngine.Rendering.Universal.Light2D>().intensity = 3;
-    }
-
-    public void CloseDoor() {
-        animator.SetBool("IsUnlocked", false);
-        isOpen = false;
-        AudioControl.Instance.PlaySFX("Door Opening", gameObject, 0.2f, 1f);
-    }
-
-    public void FlipDoor() {
-        if (isOpen) {
-            CloseDoor();
-            isOpen = false;
-        } else {
-            OpenDoor();
-            isOpen = true;
-        }
     }
 }
