+++ Assets/DoorData.cs
@@ -1,0 +1,5 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "Door Data", menuName = "Item Data/Door Data")]
+
+public class DoorData : CollectibleItem { }
