+++ Assets/DoorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a9031d5d3b2c02488b424603664e6f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
