+++ Assets/DoorPopupManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 686df78f605fe99428b0e782ed9e1fc6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
