--- Assets/Scenes/Joseph Seo/M3.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: dc85c9a3555e59a4abb871a07283c63f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
