--- Assets/Scenes/Joseph Seo/PType.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PType : MonoBehaviour
-{
-    // Start is called before the first frame update
-
-    [SerializeField] public string type;
-
-    public string getType() {
-        return type;
-    }
-
-
-
-}
