--- Assets/Scenes/Joseph Seo/PType.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 44b1ee7757ad34040b10d9a45b83e84f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
